priv/python/requirements.txt
Pillow==7.0.0
pypng==0.0.20
PyQRCode==1.2.1
pyzbar==0.1.8
